(写编码器模块:完善can通信的函数实现|修改angle_process编码值转角度的参数)

+++ components/support/angle_process.h
@@ -11,7 +11,7 @@
 
 #define NORMALIZE_TO_2PI(angle) (angle+PI) // ·¶Î§-PI~PI×ª»¯Îª0~2PI
 
-fp32 ecd_to_angle(int16_t ecd,int16_t max_ecd, int32_t offset_ecd , fp32 offset_angle);
+fp32 ecd_to_angle(int32_t ecd,int32_t max_ecd, int32_t offset_ecd , fp32 offset_angle);
 fp32 angle_normalize(fp32 angle,fp32 offset);
 
 #endif
